# Add working metadata caching