--- a/src/Spring/Spring.Aop/Aop/Framework/AutoProxy/InheritanceBasedAopConfigurer.cs
+++ b/src/Spring/Spring.Aop/Aop/Framework/AutoProxy/InheritanceBasedAopConfigurer.cs
@@ -1,5 +1,5 @@
 /*
- * Copyright © 2002-2011 the original author or authors.
+ * Copyright ï¿½ 2002-2011 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -155,7 +155,7 @@ namespace Spring.Aop.Framework.AutoProxy
         /// </remarks>
         public void PostProcessObjectFactory(IConfigurableListableObjectFactory factory)
         {
-            IList<string> objectDefinitionNames = factory.GetObjectDefinitionNames();
+            var objectDefinitionNames = factory.GetObjectDefinitionNames();
             for (int i = 0; i < objectDefinitionNames.Count; ++i)
             {
                 string name = objectDefinitionNames[i];